# public/docs/openapi.yaml
openapi: 3.0.1
info:
  title: BiciMarket Payments API
  version: 1.0.0
  description: |
    **Owner**: Rocco Paoloni · **Clerk**: `payments.bicimarket` (admin-only)

    API para gestionar pagos, reembolsos, liquidaciones y payouts.

    **Importante**: Buyers y sellers no se loguean en Payments App. Las vistas administrativas usan JWT con `publicMetadata.admin=true`.
    Los endpoints server-to-server son consumidos por otras apps con `X-Service-Token`.

    **Secciones**:
    - **Pagos**: POST/GET /payments, refund, cancel, confirm (admin override)
    - **Reembolsos**: POST/GET /refunds, detail + status update (admin only)
    - **Comprobantes**: GET /receipts, POST /receipts (Buyer App), detail
    - **Liquidaciones**: GET /settlements, mark_paid (admin), detail
    - **Payouts**: POST/GET /payouts, detail + complete (admin only)
    - **Webhooks**: Mercado Pago + shipment-delivered trigger

  x-api-specs:
    base_path: /api/v1
    webhook_path: /webhooks/mercadopago
    version: v1
    next_version: v2
servers:
  - url: http://localhost:3000
    description: Local Development
    x-env: development
  - url: https://payments.bicimarket.com
    description: Production
    x-env: production
paths:
...
  /api/v1/settlements:
    get:
      summary: Listar liquidaciones
      description: |
        **Listar liquidaciones (settlements)** por seller con paginación.

        Usado por Seller App para mostrar "Mis liquidaciones" y Admin UI.

        Querystring: `?sellerId=slp_01H…&status=paid&from=2026-04-01&to=2026-04-30`
      tags:
        - "Liquidaciones"
      security:
        - X-Service-Token: []
        - bearerAuth: []
      parameters:
        - in: query
          name: paymentId
          schema: { type: string }
          description: Filter by payment ID
        - in: query
          name: sellerId
          schema: { type: string }
          description: Filter by seller profile ID (slp_...)
        - in: query
          name: status
          schema: { type: string, enum: [pending, paid, failed, manual_review] }
        - in: query
          name: from
          schema: { type: string, format: date-time }
          description: Filter by start date (ISO 8601)
        - in: query
          name: to
          schema: { type: string, format: date-time }
          description: Filter by end date (ISO 8601)
        - in: query
          name: q
          schema: { type: string }
          description: Full-text search across id, payment_id, seller_profile_id, order_id
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
        - in: query
          name: sort
          schema: { type: string, default: '-created_at' }
          description: Sort field, prefix with `-` for descending
      responses:
        '200':
          description: Paginated list of settlements
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Settlement'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token for UI/authenticated endpoints
    X-Service-Token:
      type: apiKey
      in: header
      name: X-Service-Token
      description: Service token for inter-app communication (Buyer, Seller, Shipping)
  schemas:
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean
        next_cursor:
          type: string
          nullable: true
...
    Settlement:
      type: object
      properties:
        id:
          type: string
          description: Settlement ID (set_...)
        payment_id:
          type: string
        order_id:
          type: string
        order_seller_group_id:
          type: string
        seller_profile_id:
          type: string
        gross_amount_cents:
          type: integer
          description: Seller subtotal + shipping
        fee_amount_cents:
          type: integer
          description: Marketplace commission (10%)
        net_amount_cents:
          type: integer
          description: gross - fee
        currency:
          type: string
          default: ARS
        status:
          type: string
          enum: [pending, paid, failed, manual_review]
        paid_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...